% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EMPTY}
\alias{EMPTY}
\title{Constant for the empty square. It's value is the character "_".}
\format{An object of class \code{character} of length 1.}
\usage{
EMPTY
}
\description{
It's value is the character "_".
}
\keyword{datasets}

