% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt3.R
\name{firstAvailableMovePlayer}
\alias{firstAvailableMovePlayer}
\title{Player that always takes the first move in the list}
\usage{
firstAvailableMovePlayer(gameState)
}
\arguments{
\item{gameState}{The current gamestate}
}
\description{
Player that always takes the first move in the list
}
\examples{
finalGameState <- playGame(px,py)
}

