% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMoves}
\alias{getMoves}
\title{Get the list of valid move from the game state}
\usage{
getMoves(gameState)
}
\arguments{
\item{gameState}{The game state for which moves must be calculated}
}
\description{
Get the list of valid move from the game state
}
\examples{
gameState <- startGame()
moves <- getMoves(gameState)
}

