% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NONE}
\alias{NONE}
\title{Constant for no winner. It's value is the character "_".}
\format{An object of class \code{character} of length 1.}
\usage{
NONE
}
\description{
It's value is the character "_".
}
\keyword{datasets}

