% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt3.R
\name{playGame}
\alias{playGame}
\title{Play a game using the two strageries}
\usage{
playGame(px, po)
}
\arguments{
\item{px}{The X player strategy}

\item{po}{The O player strategy}
}
\description{
Play a game using the two strageries
}
\examples{
finalGameState <- playGame(px,py)
}

