% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{randomMovePlayer}
\alias{randomMovePlayer}
\title{Player that picks a random move}
\usage{
randomMovePlayer(gameState)
}
\arguments{
\item{gameState}{The current gamestate}
}
\description{
Player that picks a random move
}
\examples{
gameState <- startGame()
move <- randomMovePlayer(gameState)
}

