% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt3.R
\docType{package}
\name{rt3}
\alias{rt3}
\alias{rt3-package}
\title{rt3: A package to facilitate playing Tic-Tac-Toe or Naughts and Crosses.}
\description{
The rt3 package impliments functions to allow a user to simulate Tic-Tac-Toe
games. It provides a convinient gameState objescta as well as simple interface
for developing new types of players.
}
\section{rt3 functions}{

playGame

startGame
getMoves
makeMove
getWinner

randomMovePlayer
firstAvailableMovePlayer
}

