% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{startGame}
\alias{startGame}
\title{Start a new game}
\usage{
startGame()
}
\value{
gameState
}
\description{
This function starts a new game. It randomly assignes a starting user.
}
\examples{
gameState <- startGame()
}

