% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt3.R
\docType{data}
\name{X}
\alias{X}
\title{This is a R Tic-Tac-Toe library}
\format{An object of class \code{character} of length 1.}
\usage{
X
}
\description{
This is a R Tic-Tac-Toe library
}
\author{
Johan Jordaan

Constant for X player
}
\keyword{datasets}

