% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{firstAvailableMovePlayer}
\alias{firstAvailableMovePlayer}
\title{Player that always takes the first move in the list of valid moves.}
\usage{
firstAvailableMovePlayer(gameState)
}
\arguments{
\item{\link[rt3]{gameState}}{The game state that the player should act on.}
}
\description{
Internally this player calls \link[rt3]{getMoves} and then picks the first entry in the list of moves.
}
\examples{
gameState <- startGame()
move <- firstAvailableMovePlayer(gameState)
}

