% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{getMoves}
\alias{getMoves}
\title{Get the current valid moves}
\usage{
getMoves(gameState)
}
\arguments{
\item{gameState}{This is a gamestate greated by startGame}
}
\description{
Get the current valid moves
}
\examples{
gameState <- startGame()
moves <- getMoves(gameState)
}

