% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMoves}
\alias{getMoves}
\title{Get the list of valid move from the game state.}
\usage{
getMoves(gameState)
}
\arguments{
\item{gameState}{The \link[rt3]{gameState} for which moves must be calculated.}
}
\value{
validMoves An array (["integer"]) of valid moves based on the provided game state.
}
\description{
Get the list of valid move from the game state.
}
\examples{
gameState <- startGame()
validMoves <- getMoves(gameState)
}

