% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{makeMove}
\alias{makeMove}
\title{Apply the move to the current game state an produce a new game state}
\usage{
makeMove(gameState, move)
}
\arguments{
\item{gameState}{The gamestate to apply the move to}

\item{move}{One of the moves returned by getMoves()}
}
\description{
Apply the move to the current game state an produce a new game state
}
\examples{
gameState <- startGame()
gameState <- makeMove(gameState,1)
}

