% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{playGame}
\alias{playGame}
\title{Play a game of Tic-Tac-Toe using the two provided stragies.}
\usage{
playGame(px, po)
}
\arguments{
\item{px}{The X player strategy.}

\item{po}{The O player strategy.}
}
\value{
gameState The final \link[rt3]{gameState} after playing a full game.
}
\description{
Play a game of Tic-Tac-Toe using the two provided stragies.
}
\examples{
px <- firstAvailableMovePlayer
py <- randomMovePlayer
finalGameState <- playGame(px,py)
}

