% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt3.R
\name{randomMovePlayer}
\alias{randomMovePlayer}
\title{Player that picks a random move}
\usage{
randomMovePlayer(gameState)
}
\arguments{
\item{gameState}{The current gamestate}
}
\description{
Player that picks a random move
}
\examples{
finalGameState <- playGame(px,py)
}

