% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{randomMovePlayer}
\alias{randomMovePlayer}
\title{Player that picks a random move}
\usage{
randomMovePlayer(gameState)
}
\arguments{
\item{\link[rt3]{gameState}}{The game state that the player should act on.}
}
\description{
Internally this player calls \link[rt3]{getMoves} and then picks an entry in the list of moves at random.
}
\examples{
gameState <- startGame()
move <- randomMovePlayer(gameState)
}

