% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{randomMovePlayer}
\alias{randomMovePlayer}
\title{Player that picks a random move}
\usage{
randomMovePlayer(gameState)
}
\arguments{
\item{gameState}{The \link[rt3]{gameState} that the player should act on.}
}
\value{
moveIndex Index to a valid move as returned by the \link[rt3]{getMoves} function.
}
\description{
Internally this player calls \link[rt3]{getMoves} and then picks an entry
  in the list of moves at random.

  A player is a function that takes a game state as input and returns a valid move index.
}
\examples{
gameState <- startGame()
move <- randomMovePlayer(gameState)
}

