% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{startGame}
\alias{startGame}
\title{Start a new game}
\usage{
startGame()
}
\value{
\link[rt3]{gameState} A new game state.
}
\description{
This function starts a new game. It randomly assigns a starting user.
 and return a new game state object.
}
\examples{
gameState <- startGame()
}

