% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{startGame}
\alias{startGame}
\title{Start a new game}
\usage{
startGame()
}
\value{
gameState A new \link[rt3]{gameState}.
}
\description{
This function starts a new game. It randomly assigns a
 starting player and returns a new game state object.
}
\examples{
gameState <- startGame()
}

